+++ packages/react-shortcuts/src/Shortcut/index.ts
@@ -1,5 +1,6 @@
 import Shortcut from './Shortcut';
 
 export {default as useShortcut} from './hooks';
+export type {DefaultIgnoredTag} from './hooks';
 export * from './Shortcut';
 export default Shortcut;